requirements/base.txt
Babel==2.3.4
Django==1.9.6
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.2.0
Pygments==2.1.3
Sphinx==1.4.2
Unidecode==0.04.19
Werkzeug==0.11.9
alabaster==0.7.8
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
arrow==0.7.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
binaryornot==0.4.0
celery==3.1.23
chardet==2.3.0
click==6.6
cookiecutter==1.4.0
coverage==4.0.3
decorator==4.0.9
django-allauth==0.25.2
django-autoslug==1.9.3
django-braces==1.8.1
django-coverage-plugin==1.3
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-environ==0.4.0
django-extensions==1.6.7
django-floppyforms==1.6.2
django-model-utils==2.5
django-redis==4.4.2
django-secure==1.0.1
django-test-plus==1.0.12
djangorestframework==3.3.3
docutils==0.12
factory-boy==2.7.0
fake-factory==0.5.7
future==0.15.2
imagesize==0.7.1
ipaddress==1.0.16
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
jinja2-time==0.1.0
kombu==3.0.35
oauthlib==1.1.1
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
poyo==0.4.0
psycopg2==2.6.1
ptyprocess==0.5.1
py==1.4.31
pytest==2.9.1
pytest-django==2.9.1
pytest-sugar==0.7.1
python-dateutil==2.5.3
python-openid==2.2.5
pytz==2016.4
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
sqlparse==0.1.19
termcolor==1.1.0
traitlets==4.2.1
unicode-slugify==0.1.3
wheel==0.29.0
whichcraft==0.4.0
wsgiref==0.1.2
